Got rid of the helper, hoping to fix data generation on DEV
